+++ tiled/server/core.py
@@ -9,26 +9,20 @@ import sys
 import uuid
 from collections import defaultdict
 from datetime import datetime, timedelta
-from functools import lru_cache
 from hashlib import md5
-from typing import Any, Optional
+from typing import Any
 
 import dateutil.tz
 import jmespath
 import msgpack
 import orjson
-import pydantic
-from fastapi import Depends, HTTPException, Query, Request, Response
+from fastapi import HTTPException, Response
 from starlette.responses import JSONResponse, Send, StreamingResponse
 
 # Some are not directly used, but they register things on import.
 from .. import queries
 from ..adapters.mapping import MapAdapter
-from ..media_type_registration import (
-    serialization_registry as default_serialization_registry,
-)
 from ..queries import KeyLookup, QueryValueError
-from ..query_registration import query_registry as default_query_registry
 from ..structures import node  # noqa: F401
 from ..structures.dataframe import serialize_arrow
 from ..utils import (
@@ -38,7 +32,6 @@ from ..utils import (
     modules_available,
 )
 from . import models
-from .authentication import get_current_principal
 from .etag import tokenize
 from .utils import record_timing
 
@@ -59,92 +52,6 @@ if modules_available("xarray"):
 
 _FILTER_PARAM_PATTERN = re.compile(r"filter___(?P<name>.*)___(?P<field>[^\d\W][\w\d]+)")
 _LOCAL_TZINFO = dateutil.tz.gettz()
-
-
-@lru_cache(1)
-def get_query_registry():
-    "This may be overridden via dependency_overrides."
-    return default_query_registry
-
-
-@lru_cache(1)
-def get_serialization_registry():
-    "This may be overridden via dependency_overrides."
-    return default_serialization_registry
-
-
-def get_root_tree():
-    raise NotImplementedError(
-        "This should be overridden via dependency_overrides. "
-        "See tiled.server.app.build_app()."
-    )
-
-
-def entry(
-    path: str,
-    request: Request,
-    principal: str = Depends(get_current_principal),
-    root_tree: pydantic.BaseSettings = Depends(get_root_tree),
-):
-    path_parts = [segment for segment in path.split("/") if segment]
-    entry = root_tree.authenticated_as(principal)
-    try:
-        # Traverse into sub-tree(s).
-        for segment in path_parts:
-            try:
-                unauthenticated_entry = entry[segment]
-            except (KeyError, TypeError):
-                raise NoEntry(path_parts)
-            if hasattr(unauthenticated_entry, "authenticated_as"):
-                with record_timing(request.state.metrics, "acl"):
-                    entry = unauthenticated_entry.authenticated_as(principal)
-            else:
-                entry = unauthenticated_entry
-        return entry
-    except NoEntry:
-        raise HTTPException(status_code=404, detail=f"No such entry: {path_parts}")
-
-
-def block(
-    # Ellipsis as the "default" tells FastAPI to make this parameter required.
-    block: str = Query(..., regex="^[0-9]*(,[0-9]+)*$"),
-):
-    "Specify and parse a block index parameter."
-    if not block:
-        return ()
-    return tuple(map(int, block.split(",")))
-
-
-def expected_shape(
-    expected_shape: Optional[str] = Query(
-        None, min_length=1, regex="^[0-9]+(,[0-9]+)*$|^scalar$"
-    ),
-):
-    "Specify and parse an expected_shape parameter."
-    if expected_shape is None:
-        return
-    if expected_shape == "scalar":
-        return ()
-    return tuple(map(int, expected_shape.split(",")))
-
-
-def slice_(
-    slice: str = Query(None, regex="^[0-9,:]*$"),
-):
-    "Specify and parse a block index parameter."
-    import numpy
-
-    # IMPORTANT We are eval-ing a user-provider string here so we need to be
-    # very careful about locking down what can be in it. The regex above
-    # excludes any letters or operators, so it is not possible to execute
-    # functions or expensive arithmetic.
-    return tuple(
-        [
-            eval(f"numpy.s_[{dim!s}]", {"numpy": numpy})
-            for dim in (slice or "").split(",")
-            if dim
-        ]
-    )
 
 
 def len_or_approx(tree):
@@ -653,7 +560,8 @@ def resolve_media_type(request):
     return media_type
 
 
-def json_or_msgpack(request, content, media_type, expires=None, headers=None):
+def json_or_msgpack(request, content, expires=None, headers=None):
+    media_type = resolve_media_type(request)
     content_as_dict = content.dict()
     with record_timing(request.state.metrics, "tok"):
         etag = md5(str(content_as_dict).encode()).hexdigest()
